fix: apply transport metadata at dequeue time, extract shared buildTransportHints
